added signIn to userControllers
